Reimplement IPC backend
This commit reimplements the IPC abstraction that's used to simplify
the process of sending type-safe messages between processes while
accounting for size limitations of IPC messages.

The previous implementation used Messengers, which added complexity and
was very restrictive about the size of the data that can be sent in the
message. It was so limited, in fact, that the previous implementation
would crash when sending payloads that were too large, meaning that some
of the basic functionality never worked (and indicates a lack of testing
on my part 😶).

The new implementation directly uses Binders, which is effectively as
low-level as you can get with IPC. The overhead has been greatly reduced
with this implementation, and instrumentation tests now validate the
correctness of this implementation.