Rework base Encore media models
This commit overhauls the basic QueueItem and MediaItem models for more
flexibility for library consumers. The MediaItem model itself is now an
interface with fewer opinions on what fields it has. This lets consumers
have much more control over what data is contained in these objects. One
major side-effect of this is that virtually every Encore playback class
has to be generic over a given type of MediaItem. The library classes
themselves are completely agnostic to this data, hence the unbound
types in a lot of places, and the end result is a lot of noise in adding
generic types that look like `<M : MediaItem>` to virtually every
playback-related class.