Rename MediaItem to MediaObject
This rename is to avoid a conflict with the androidx model of the same
name. This is helpful both from the library perspective as well as from
an app perspective. From the perspective of the Encore libraries, this
avoids a scenario where it's required to use the fully-qualified name
for one of the classes, since it's not absurd for them both to appear in
the same file. For consumers, this rename will fix incorrect imports
since Encore needs to expose the androidx media library and therefore
androidx's MediaItem will always be available to consumers, even though
they have no need for it.